refector: separate Component
