+++ src/components/TitleText.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const HeaderText = styled.h1`
+  font-size: 2.5rem; /* 제목 폰트 사이즈 */
+  text-align: left; /* 왼쪽 정렬 */
+`;
+
+const HighlightText = styled.span`
+  color: #6e6e73; /* 강조 텍스트 색상 */
+`;
+
+export { HeaderText, HighlightText };
