+++ src/pages/MainPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import Header from '@components/Header';
 import BrandSection from '@components/BrandSection';
-import Containers from '@components/Containers';
+import TravelRecommendations from '@components/TravelRecommendations';
 
 const MainContainer = styled.div`
   display: flex;
@@ -11,16 +11,12 @@ const MainContainer = styled.div`
   width: 100%;
 `;
 
-const ContainerWrapper = styled.div`
-  display: flex;
-  gap: 1.5cm; /* 초기 간격 */
-`;
-
 const MainPage = () => {
   return (
     <MainContainer>
       <Header />
-      <Containers />
+      <BrandSection />
+      <TravelRecommendations />
     </MainContainer>
   );
 };
