+++ src/pages/index.jsx
@@ -1,0 +1,12 @@
+
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+import './index.css'; // 필요에 따라 CSS 파일을 임포트하세요
+
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
+);
