+++ src/styles/MainPage.style.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const MainContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+`;
+
+export const ContainerWrapper = styled.div`
+  display: flex;
+  gap: 1.5cm; /* 초기 간격 */
+`;
